refactor: extract ArtifactHandler from MCPServerV2
Move 6 artifact management methods into dedicated ArtifactHandler.
Move _test_artifacts_base to handler; update test references.
