+++ src/llm_orc/mcp/server.py
@@ -18,6 +18,7 @@ from mcp.server.fastmcp import Context, FastMCP
 from llm_orc.core.config.config_manager import ConfigurationManager
 from llm_orc.core.config.ensemble_config import EnsembleLoader, assert_no_cycles
 from llm_orc.core.execution.artifact_manager import ArtifactManager
+from llm_orc.mcp.handlers.artifact_handler import ArtifactHandler
 from llm_orc.mcp.handlers.help_handler import HelpHandler
 from llm_orc.mcp.handlers.profile_handler import ProfileHandler
 
@@ -113,7 +114,6 @@ class MCPServerV2:
     # Optional test overrides for directories
     _test_library_dir: Path | None = None
     _test_scripts_dir: Path | None = None
-    _test_artifacts_base: Path | None = None
     _test_ollama_status: dict[str, Any] | None = None
 
     def __init__(
@@ -134,6 +134,7 @@ class MCPServerV2:
         self._executor = executor  # Lazily created if None
         self._help_handler = HelpHandler()
         self._profile_handler = ProfileHandler(self.config_manager)
+        self._artifact_handler = ArtifactHandler()
         self._mcp = FastMCP("llm-orc")
         self._setup_resources()
         self._setup_tools()
@@ -2095,137 +2096,15 @@ class MCPServerV2:
     # Artifact Management Tool Implementations
     # =========================================================================
 
-    def _get_artifacts_base(self) -> Path:
-        """Get artifacts base directory."""
-        if self._test_artifacts_base is not None:
-            return self._test_artifacts_base
-        return Path.cwd() / ".llm-orc" / "artifacts"
-
     async def _delete_artifact_tool(self, arguments: dict[str, Any]) -> dict[str, Any]:
-        """Delete an artifact.
-
-        Args:
-            arguments: Tool arguments.
-
-        Returns:
-            Deletion result.
-        """
-        import shutil
-
-        artifact_id = arguments.get("artifact_id")
-        confirm = arguments.get("confirm", False)
-
-        if not artifact_id:
-            raise ValueError("artifact_id is required")
-        if not confirm:
-            raise ValueError("Confirmation required to delete artifact")
-
-        # Parse artifact_id (format: ensemble/timestamp)
-        parts = artifact_id.split("/")
-        if len(parts) != 2:
-            raise ValueError("Invalid artifact_id format (expected ensemble/timestamp)")
-
-        ensemble_name, timestamp = parts
-
-        # Find artifact directory
-        artifacts_base = self._get_artifacts_base()
-        artifact_dir = artifacts_base / ensemble_name / timestamp
-
-        if not artifact_dir.exists():
-            raise ValueError(f"Artifact '{artifact_id}' not found")
-
-        # Delete
-        shutil.rmtree(artifact_dir)
-
-        return {"deleted": True, "artifact_id": artifact_id}
+        """Delete an artifact."""
+        return await self._artifact_handler.delete_artifact(arguments)
 
     async def _cleanup_artifacts_tool(
         self, arguments: dict[str, Any]
     ) -> dict[str, Any]:
-        """Cleanup old artifacts.
-
-        Args:
-            arguments: Tool arguments.
-
-        Returns:
-            Cleanup result.
-        """
-        import time
-
-        ensemble_name = arguments.get("ensemble_name")
-        older_than_days = arguments.get("older_than_days", 30)
-        dry_run = arguments.get("dry_run", True)
-
-        artifacts_base = self._get_artifacts_base()
-        cutoff_time = time.time() - (older_than_days * 24 * 60 * 60)
-
-        ensemble_dirs = self._get_ensemble_artifact_dirs(artifacts_base, ensemble_name)
-        would_delete, deleted = self._process_old_artifacts(
-            ensemble_dirs, cutoff_time, dry_run
-        )
-
-        if dry_run:
-            return {
-                "dry_run": True,
-                "would_delete": would_delete,
-                "count": len(would_delete),
-            }
-        return {"dry_run": False, "deleted": deleted, "count": len(deleted)}
-
-    def _get_ensemble_artifact_dirs(
-        self, artifacts_base: Path, ensemble_name: str | None
-    ) -> list[Path]:
-        """Get list of ensemble artifact directories to check.
-
-        Args:
-            artifacts_base: Base artifacts directory.
-            ensemble_name: Optional specific ensemble to check.
-
-        Returns:
-            List of ensemble directories.
-        """
-        if ensemble_name:
-            return [artifacts_base / ensemble_name]
-        if artifacts_base.exists():
-            return [d for d in artifacts_base.iterdir() if d.is_dir()]
-        return []
-
-    def _process_old_artifacts(
-        self, ensemble_dirs: list[Path], cutoff_time: float, dry_run: bool
-    ) -> tuple[list[str], list[str]]:
-        """Process and optionally delete old artifacts.
-
-        Args:
-            ensemble_dirs: List of ensemble directories to check.
-            cutoff_time: Unix timestamp cutoff for deletion.
-            dry_run: If True, don't actually delete.
-
-        Returns:
-            Tuple of (would_delete list, deleted list).
-        """
-        import shutil
-
-        would_delete: list[str] = []
-        deleted: list[str] = []
-
-        for ensemble_dir in ensemble_dirs:
-            if not ensemble_dir.exists():
-                continue
-
-            for artifact_dir in ensemble_dir.iterdir():
-                if not artifact_dir.is_dir():
-                    continue
-
-                mtime = artifact_dir.stat().st_mtime
-                if mtime < cutoff_time:
-                    artifact_id = f"{ensemble_dir.name}/{artifact_dir.name}"
-                    would_delete.append(artifact_id)
-
-                    if not dry_run:
-                        shutil.rmtree(artifact_dir)
-                        deleted.append(artifact_id)
-
-        return would_delete, deleted
+        """Cleanup old artifacts."""
+        return await self._artifact_handler.cleanup_artifacts(arguments)
 
     # =========================================================================
     # Script Management Tool Implementations (Low Priority)
